refactor(extractor): moving Phind-Search sources extracting logic
