+++ src/scripts/content/extractor/types/ExtractorPhindSearch.js
@@ -2,7 +2,8 @@ import {clickElements} from "../../interact/interact";
 import {capitalizeFirst, formatLineBreaks} from "../../../shared/formatter/formatText";
 import {setFileHeader} from "../../../shared/formatter/formatMarkdown";
 import ExtractorPhind from "./ExtractorPhind";
-import {safeExecute, sleep} from "../../../shared/utils/jsShorteners";
+import {safeExecute} from "../../../shared/utils/jsShorteners";
+import ExtractorSourcesPhindSearch from "../sources/ExtractorSourcesPhindSearch";
 
 
 export default class ExtractorPhindSearch extends ExtractorPhind {
@@ -31,23 +32,6 @@ export default class ExtractorPhindSearch extends ExtractorPhind {
 
       const selectPagination = content.querySelectorAll('.pagination button');
 
-      // Create formatted document for each answer message
-      async function fetchSources(selectPagination) {
-        let res = "";
-        let i = 1;
-        for (const elt of selectPagination) {
-          elt.click();
-          await sleep(1);
-          const selectSources = document.querySelectorAll('a.mb-0');
-          selectSources.forEach((sourceElt) => {
-            res += "\n- " + format(sourceElt.outerHTML).replace("[", `[(${i}) `);
-            i++;
-          });
-        }
-        selectPagination[0] && selectPagination[0].click();
-        return res;
-      }
-
       async function generateMessageText(selectUserQuestion, selectAiAnswer, selectAiModel, selectPagination) {
         const userPart = `\n## User\n` + format(formatLineBreaks(selectUserQuestion.innerHTML)).replace("  \n", "") + '\n';
 
@@ -61,7 +45,7 @@ export default class ExtractorPhindSearch extends ExtractorPhind {
         const aiPart = `\n` + aiIndicator + aiAnswer.substring(index + 2);
 
         const paginationPart = selectPagination.length > 0
-          ? `\n\n---\n**Sources:**` + await safeExecute(await fetchSources(selectPagination)) + "\n\n"
+          ? `\n\n---\n**Sources:**` + await safeExecute(await new ExtractorSourcesPhindSearch().extractSources(selectPagination, format)) + "\n\n"
           : "";
 
         return userPart + aiPart + paginationPart;
